Introduce new plugin property `lookup_groups_base`.
An additional possibility to define a separate groups_base_dn for the group lookup.
This allow us to set a restricted groups_base, which does not include all nested groups. Because the plugin considers now also group memberships outside the restricted groups_base and uses the lookup_groups_base if it exists.