Handle CNs with commas in them in group membership resolution:
When splitting DNs on commas in getGroupsForPrincipal() we need
to take care to mask commas that are part of CNs beforehand, and
unmask them again after splitting.

In addition, it seems that either delegate.search() or possibly
the Python ldap module already handles proper escaping of commas
in CNs when used in filters. This means that in order to filter
for a CN containing a comma, the comma in the CN should just be
included verbatim, *without* any blackslash before the comma.

We therefore remove the backslash as well when unmasking.